pyproject.toml

[project]
name = "nd-blueteam-2"
version = "0.1.0"
description = ""
authors = [
    {name = "psaboia", email = "psaboiawork@gmail.com"},
]
dependencies = [
    "adversarial-robustness-toolbox==1.15.1",
    "charset-normalizer==3.1.0",
    "cmake==3.26.1",
    "contourpy==1.0.7",
    "cycler==0.11.0",
    "filelock==3.11.0",
    "fonttools==4.39.3",
    "idna==3.4",
    "Jinja2==3.1.2",
    "joblib==1.3.2",
    "kiwisolver==1.4.4",
    "lit==16.0.0",
    "MarkupSafe==2.1.2",
    "matplotlib==3.7.1",
    "mpmath==1.3.0",
    "networkx==3.1",
    "numpy==1.24.2",
    "nvidia-cublas-cu11==11.10.3.66",
    "nvidia-cuda-cupti-cu11==11.7.101",
    "nvidia-cuda-nvrtc-cu11==11.7.99",
    "nvidia-cuda-runtime-cu11==11.7.99",
    "nvidia-cudnn-cu11==8.5.0.96",
    "nvidia-cufft-cu11==10.9.0.58",
    "nvidia-curand-cu11==10.2.10.91",
    "nvidia-cusolver-cu11==11.4.0.1",
    "nvidia-cusparse-cu11==11.7.4.91",
    "nvidia-nccl-cu11==2.14.3",
    "nvidia-nvtx-cu11==11.7.91",
    "packaging==23.0",
    "pandas==2.0.3",
    "Pillow==9.5.0",
    "pycocotools==2.0.6",
    "pyparsing==3.0.9",
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "requests==2.28.2",
    "scikit-learn==1.1.3",
    "scipy==1.11.2",
    "six==1.16.0",
    "sympy==1.11.1",
    "threadpoolctl==3.2.0",
    "torch==2.0.0",
    "torchaudio==2.0.1",
    "torchvision==0.15.1",
    "tqdm==4.66.1",
    "triton==2.0.0",
    "typing-extensions==4.5.0",
    "tzdata==2023.3",
    "urllib3==1.26.15",
    "opencv-python>=4.8.0.76",
    "pyyaml>=6.0.1",
]
requires-python = ">=3.10,<3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

